UtilityのSplitとYougineMathをユーザと共有するプロジェクトに入れた
