+++ Engine/Core/src/yougine/Editor/ComponentViewer.h
@@ -7,8 +7,8 @@
 #include "ComponentsIncluder.h"
 #include <string>
 #include <any>
-#include "../utilitys/YougineMath.h"
-#include "../utilitys/Split.cpp"
+#include "UserShare/utilitys/YougineMath.h"
+#include "UserShare/utilitys/Split.cpp"
 #include "../utilitys/Quaternion.h"
 
 #include "PropertiesInputField.h"
