+++ Engine/Core/src/yougine/Editor/ProjectWindows/Assets/element/Model/Material/ShaderInputParameters/ShaderInputAndTypeStruct.h
@@ -3,7 +3,7 @@
 #include <string>
 
 #include "../ShaderInputParameterType.h"
-#include "../../../../../../../utilitys/YougineMath.h"
+#include "UserShare/utilitys/YougineMath.h"
 #include "../../../../../../../utilitys/view/parameters/AssetReference.h"
 
 namespace editor::projectwindows::assets::elements::model::materials::shaderinputparameters
