+++ Engine/Core/src/yougine/Editor/PropertiesInputField.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <memory>
 #include <string>
-#include "../utilitys/YougineMath.h"
-#include "../utilitys/Split.cpp"
+#include "UserShare/utilitys/YougineMath.h"
+#include "UserShare/utilitys/Split.cpp"
 #include "../utilitys/Quaternion.h"
 #include <any>
 #include "imgui/imgui.h"
