+++ Engine/Core/src/yougine/Editor/ShaderGraph/ShaderGraphWindow.h
@@ -5,7 +5,7 @@
 #include "Nodes/ShaderGraphNode.h"
 #include "Nodes/ShaderGraphInputNodes.h"
 #include "Nodes/UnlitShaderGraphNode.h"
-#include "../../utilitys/Split.cpp"
+#include "UserShare/utilitys/Split.cpp"
 #include "ShaderfileOverwriter.h"
 #include "ShaderGraphInputFieldViewer.h"
